pyproject.toml
[project]
name = "twin"
version = "0.1.0"
description = "A high-performance FastAPI application providing a streaming chat interface using the Google Gemini AI SDK."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.61.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
pythonpath = "."
